+++ lib/simple_crud.rb
@@ -1,7 +1,7 @@
 require_relative 'simple_crud/simple_crud_controller'
 
 module SimpleCrud
-  VERSION = '0.2'.freeze
+  VERSION = '0.5'.freeze
   def self.configure
     yield Config
   end
